Dependency encoders: make the root relation label configurable
The root label used by post-processing was hardcoded to be
`ROOT`. However, this is incorrect for various schemes, including
universal dependencies. This change makes the root relation
configurable per encoder/decoder.

The part-of-speech tag of the root node is still fixed to `ROOT`. This
tag is only used internally and changing this would break existing
label files.

This issue was reported by @twuebi.